pyproject.toml
[project]
name = "milton"
version = "0.1.0"
description = ""
authors = [
    {name = "cyberphor"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "openai (>=1.61.0,<2.0.0)",
    "autogen-agentchat (>=0.2,<0.4)",
    "streamlit (>=1.41.1,<2.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
